fixed CG errors end find errors on compilation

+++ taskaev.matvei/P3/formingString.hpp
@@ -4,7 +4,7 @@
 
 namespace taskaev
 {
-  char* formingString(const char* string, size_t& size, size_t newSize);
+  char* formingString(const char* string, int& size, int newSize);
 }
 #endif
